pyproject.toml
[project]
name = "mlops-course1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.3",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
